fix : intersection observer 커스텀 훅으로 분리
